enable dynamic AA
