+++ Project.toml
@@ -7,6 +7,7 @@ version = "0.0.2"
 ACE1 = "e3f9bc04-086e-409a-ba78-e9769fe067bb"
 ACE1x = "5cc4c08c-8782-4a30-af6d-550b302e9707"
 ACEbase = "14bae519-eb20-449c-a949-9c58ed33163e"
+Bumper = "8ce10254-0962-460f-a3d8-1f77fea1446e"
 ChainRules = "082447d4-558c-5d27-93f4-14fc19e9eca2"
 ChunkSplitters = "ae650224-84b6-46f8-82ea-d812ca08434e"
 DynamicPolynomials = "7c1d4256-1411-5781-91ec-d7bc3513ac07"
@@ -25,6 +26,7 @@ SpheriCart = "5caf2b29-02d9-47a3-9434-5931c85ba645"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 StaticPolynomials = "62e018b1-6e46-5407-a5a7-97d4fbcae734"
 StrideArrays = "d1fa6d79-ef01-42a6-86c9-f7c551f8593b"
+WithAlloc = "fb1aa66a-603c-4c1d-9bc4-66947c7b08dd"
 
 [compat]
 ChunkSplitters = "2"
@@ -32,7 +34,7 @@ Folds = "0.2"
 LoopVectorization = "0.12"
 NamedTupleTools = "0.14"
 ObjectPools = "0.3"
-Polynomials4ML = "0.2.7"
+Polynomials4ML = "0.2.11"
 julia = "1"
 
 [extras]
